9 - Erweiterung Kassenbuch bearbeiten
Jede GUI in eigene Klasse um Übersichtlichkeit zu verbessern
