+++ src/main/java/de/sg/computerinsel/tools/kassenbuch/KassenbuchBearbeitenUtils.java
@@ -20,34 +20,32 @@ import de.sg.computerinsel.tools.kassenbuch.model.Rechnung;
  * @author Sita Geßner
  */
 public final class KassenbuchBearbeitenUtils {
-
-	private static final Logger LOGGER = LoggerFactory.getLogger(KassenbuchBearbeitenUtils.class);
-
-	private static final int INDEX_LINE_DATE = 0;
-
-	static final int INDEX_LINE_ART = 1;
-
-	static final int INDEX_LINE_SUM = 4;
 	
-	static final int MAX_LENGTH_LINE = 5;
+	private static final Logger LOGGER = LoggerFactory.getLogger(KassenbuchBearbeitenUtils.class);
+	
+	private static final int INDEX_LINE_DATE = 0;
+	
+	static final int INDEX_LINE_ART = 1;
+	
+	static final int INDEX_LINE_SUM = 4;
 
+	static final int MAX_LENGTH_LINE = 5;
+	
 	private KassenbuchBearbeitenUtils() {
 	}
-	
-	public static File addKassenbuchEintrag(final String filePath, final String verwendungstext, final Date datum, final BigDecimal betrag,
-	        final boolean isNegative) {
+
+	public static File addKassenbuchEintrag(final String filePath, final Rechnung rechnung) {
 		final List<Rechnung> rechnungen = readRechnungenFromCsvFile(filePath);
-		rechnungen.add(createNeueEintragung(verwendungstext, datum, betrag, isNegative));
+		rechnungen.add(rechnung);
 		final Rechnung ausgangsRechnung = rechnungen.get(0);
 		rechnungen.remove(0);
 		final File csvFile = KassenbuchErstellenUtils.createCsv(rechnungen, ausgangsRechnung,
-		        filePath.substring(0, filePath.lastIndexOf(File.separator)));
+				filePath.substring(0, filePath.lastIndexOf(File.separator)));
 		KassenbuchErstellenUtils.createPdf(rechnungen, ausgangsRechnung, filePath.substring(0, filePath.lastIndexOf(File.separator)));
 		return csvFile;
 	}
-	
-	private static Rechnung createNeueEintragung(final String verwendungstext, final Date datum, final BigDecimal betrag,
-			final boolean isNegative) {
+
+	static Rechnung createNeueEintragung(final String verwendungstext, final Date datum, final BigDecimal betrag, final boolean isNegative) {
 		final Rechnung rechnung = new Rechnung();
 		rechnung.setRechnungsbetrag(getBetrag(betrag, isNegative));
 		rechnung.setRechnungsdatum(datum);
@@ -55,7 +53,7 @@ public final class KassenbuchBearbeitenUtils {
 		LOGGER.info("Eintrag hinzugefügt: {}", rechnung);
 		return rechnung;
 	}
-	
+
 	/**
 	 * Wenn der Radiobutton <code>-</code> gesetzt wurde, muss der Betrag negativ sein werden.<br>
 	 * Wenn der Radiobutton <code>+</code> gesetzt wurde, muss der Betrag positiv sein werden.
@@ -73,7 +71,7 @@ public final class KassenbuchBearbeitenUtils {
 		}
 		return result;
 	}
-
+	
 	private static List<Rechnung> readRechnungenFromCsvFile(final String filePath) {
 		final List<Rechnung> rechnungen = new ArrayList<>();
 		String[] items = null;
@@ -85,13 +83,13 @@ public final class KassenbuchBearbeitenUtils {
 				if (items.length == MAX_LENGTH_LINE) {
 					// Gesamtbeträge und Ausgangsbeträge ignorieren, außer ersten Ausgangsbetrag
 					if (items[INDEX_LINE_ART].contains(Rechnung.GESAMTBETRAG)
-							|| (items[INDEX_LINE_ART].contains(Rechnung.AUSGANGSBETRAG) && counter > 0)) {
+					        || (items[INDEX_LINE_ART].contains(Rechnung.AUSGANGSBETRAG) && counter > 0)) {
 						LOGGER.info("Gesamtbetrag vom {}: {}", items[INDEX_LINE_DATE], items[INDEX_LINE_SUM]);
 						continue;
 					} else {
 						final Rechnung rechnung = new Rechnung();
 						rechnung.setRechnungsdatum(StringUtils.isNotBlank(items[INDEX_LINE_DATE]) ? KassenbuchErstellenUtils.DATE_FORMAT
-						        .parse(items[INDEX_LINE_DATE]) : null);
+								.parse(items[INDEX_LINE_DATE]) : null);
 						rechnung.setRechnungsnummer(StringUtils.replace(items[INDEX_LINE_ART], "Rechnung: ", ""));
 						rechnung.setRechnungsbetrag(new BigDecimal(normalizeCurrencyValue(extractBetrag(items))));
 						rechnungen.add(rechnung);
@@ -110,17 +108,17 @@ public final class KassenbuchBearbeitenUtils {
 			LOGGER.error("Fehler beim Lesen der Datei {}, {}", filePath, e.getMessage());
 		} catch (final ParseException e) {
 			LOGGER.error("Rechnungsdatum: '{}' kann nicht geparst werden: {} ", items == null ? null : items[INDEX_LINE_DATE],
-					e.getMessage());
+			        e.getMessage());
 		} catch (final NumberFormatException e) {
 			LOGGER.error("Fehler beim Lesen des Rechnungsbetrages: {}", items == null ? null : extractBetrag(items), e.getMessage());
 		}
 		return rechnungen;
 	}
-
+	
 	static String extractBetrag(final String[] items) {
 		return items[2] == null || StringUtils.isBlank(items[2]) ? items[3] : items[2];
 	}
-	
+
 	public static String normalizeCurrencyValue(final String value) {
 		return StringUtils.replaceEach(value, new String[] { "€", "EUR", "," }, new String[] { "", "", "." }).trim();
 	}
